ADC: upgrage from 8 bit resolution to 10 bit (int16_t).

+++ src/adc.h
@@ -1,6 +1,6 @@
 #ifndef ADC_H
 #define ADC_H
 
-uint8_t analogRead( uint8_t pin );
+uint16_t analogRead( int pin );
 
 #endif
